requirements.txt
# Core ML Libraries
torch>=1.7.0
torchvision>=0.8.0
pretrainedmodels>=0.7.4

# Image Processing
opencv-python>=4.5.0
Pillow>=8.0.0
scikit-image>=0.17.0

# Scientific Computing
numpy>=1.19.0
pandas>=1.2.0

# Web Framework (for demo)
gradio>=3.0.0
flask>=2.0.0
flask-cors>=3.0.0
werkzeug>=2.0.0

# Utilities
tqdm>=4.60.0

